+++ analytics/analytics/models/__init__.py
@@ -1,0 +1,16 @@
+from analytics.models.events import (
+    AttributionReferrerEvent,
+    DetailPageEvent,
+    DetailPageEvents,
+    ResultClickedEvent,
+    SearchEvent,
+    SearchRatingEvent,
+)
+from analytics.models.orm import Image
+from analytics.models.reports import (
+    AttributionRefererReport,
+    SourceUsageReport,
+    TopResultsReport,
+    TopSearchesReport,
+    UsageReport,
+)
